+++ backend/src/lib/betterauth/auth.instance.ts
@@ -1,13 +1,41 @@
 import { AuthDatabaseAdapter } from '../../adapters/auth.adapter';
+import { getRedisClient } from '../../adapters/cache.adapter';
 import { getTrustedOrigins } from '../cors';
 import { sendMagicLinkEmail } from '../email/magic-link.handler';
 
-import { createAuth } from './betterauth';
+import { createAuth, type AuthSecondaryStorage } from './betterauth';
 
 const authDb = new AuthDatabaseAdapter();
+
+/**
+ * Shared Redis secondary storage — used by Better Auth's rateLimit block so
+ * auth-endpoint throttling shares the same Redis instance as the app-level
+ * rate limiter (keyspace prefix: 'better-auth:').
+ */
+const secondaryStorage: AuthSecondaryStorage = {
+  async get(key) {
+    const redis = getRedisClient();
+    return redis.get(`better-auth:${key}`);
+  },
+  async set(key, value, ttl) {
+    const redis = getRedisClient();
+    if (ttl != null && ttl > 0) {
+      await redis.setex(`better-auth:${key}`, ttl, value);
+    } else {
+      // Better Auth occasionally writes keys without a TTL; cap at 30 days
+      // to prevent unbounded Redis growth.
+      await redis.setex(`better-auth:${key}`, 60 * 60 * 24 * 30, value);
+    }
+  },
+  async delete(key) {
+    const redis = getRedisClient();
+    await redis.del(`better-auth:${key}`);
+  },
+};
 
 export const auth = createAuth({
   authDb,
   getTrustedOrigins,
   sendMagicLinkEmail,
+  secondaryStorage,
 });
